+++ front/src/app/documents/request-in-process-view/request-in-process-view.component.ts
@@ -201,12 +201,43 @@ export class RequestInProcessViewComponent {
 	}
 
 	protected onDownload(file: FileManager) {
-		const url = `${this.http.getAppUrl()}/attachments/${file.file_path}`;
-		if (file.extension_file === 'pdf') {
-			this.documentViewerService.open(url, this.currentShipping.company_name);
-		} else {
-			this.http.openDocument(url);
-		}
+		const requestId = this.currentShipping.uuid;
+		const fileUuid = file.uuid;
+		this.mask.showBlockUI('Procesando descarga de certificado...');
+		this.issuanceService.getDownloadFile(requestId, fileUuid).subscribe({
+			next: (meta) => {
+				this.mask.hideBlockUI();
+				if (file.extension_file === 'pdf') {
+					this.documentViewerService.open(meta.download_url, this.currentShipping.company_name + ' - ' + file.file_name);
+				} else {
+					this.http.openDocument(meta.download_url);
+				}
+			},
+			error: () => {
+				this.mask.hideBlockUI();
+				this.issuanceLoading = false;
+			}
+		});
+	}
+
+	/**
+ * Obtiene la metadata de descarga del P12.
+ */
+	protected onGetDownloadMeta(): void {
+		const requestId = this.currentShipping.uuid;
+		this.issuanceLoading = true;
+		this.mask.showBlockUI('Procesando descarga de certificado...');
+		this.issuanceService.getDownloadFileUrl(requestId).subscribe({
+			next: (meta) => {
+				this.mask.hideBlockUI();
+				window.open(meta.download_url, '_blank');
+				this.issuanceLoading = false;
+			},
+			error: () => {
+				this.mask.hideBlockUI();
+				this.issuanceLoading = false;
+			}
+		});
 	}
 
 	protected selectFile(file: FileManager) {