+++ front/src/environments/environment.prod.ts
@@ -6,7 +6,7 @@ export const environment = {
   APPURL: 'https://api-certs.matias-api.com',
   APIURL: 'https://api-certs.matias-api.com/api/v1',
   WOMPI_PUBLIC_KEY: 'pub_prod_XXXXXXXXXXXXXXXX',
-  VERSION: '1.8.0',
+  VERSION: '1.9.0',
   config: {
     name: 'MATICERTS',
     title: 'MATICERTS - Aplicación para la gestión de solicitudes de certificados de firma digital',