+++ front/src/environments/environment.ts
@@ -10,7 +10,7 @@ export const environment = {
   APPURL: 'http://cm-api.test',
   APIURL: 'http://cm-api.test/api/v1',
   WOMPI_PUBLIC_KEY: 'pub_test_q4gEVnZWpzfEROZMScHdZgH4ChcGHW2E',
-  VERSION: '1.8.0',
+  VERSION: '1.9.0',
   config: {
     name: 'MATICERTS',
     title: 'MATICERTS - Aplicación para la gestión de solicitudes de certificados de firma digital',